Move constanst into new file
