+++ src/utils/constants.js
@@ -1,0 +1,10 @@
+export const ClashingProperty = {
+  startTime: 'startTime',
+  endTime: 'endTime',
+  startAndEndTime: 'startAndEndTime',
+};
+
+export const inputNames = {
+  shiftStartTime: 'shift-start-time',
+  shiftFinishTime: 'shift-finish-time',
+};
