+++ src/utils/time-entry-utils/timeEntryUtils.js
@@ -59,14 +59,3 @@ export const isFinishTimeOnNextDay = (startTimeValue, finishTimeValue) => {
   }
   return false;
 };
-
-export const ClashingProperty = {
-  startTime: 'startTime',
-  endTime: 'endTime',
-  startAndEndTime: 'startAndEndTime',
-};
-
-export const inputNames = {
-  shiftStartTime: 'shift-start-time',
-  shiftFinishTime: 'shift-finish-time',
-};
